requirements.txt
asyncio
autogen_ext==0.4.8
autogen-agentchat==0.4.8
autogen-ext[openai]
pandas>=2.0.0
openpyxl>=3.1.0
pyautogen>=0.2.0
python-dotenv>=0.19.0
pydantic==2.10.6
xlsxwriter==3.2.2
streamlit>=1.30.0
llama-index-core>=0.10.0
docling>=0.1.0
pydantic>=2.0.0
markdown==3.7